Create a version of the check that verifies the IP's "colour"
SNDS codifies the state of an IP as a colour in its web interface. Since
SNDS rarely adds a full block on an IP, it's nice to know when your
status is not green: it might mean that your server is sending out too
much emails suddenly or is sending undesirable content.

The new check was written by my colleague, Kienan Stewart.